fix ByteTensor to FloatTensor, fix train notebook WIP

+++ src/heroic_gm/data/heroic_dataset.py
@@ -16,7 +16,7 @@ class HeroicDataset(Dataset):
         self.transform = transforms.Compose([
             transforms.PILToTensor()
         ])
-        self.data['image_tensor']=self.data['image'].apply(lambda x : self.transform(Image.open(BytesIO(x['bytes']))))
+        self.data['image_tensor']=self.data['image'].apply(lambda x : self.transform(Image.open(BytesIO(x['bytes']))).type(torch.FloatTensor))
         self.data=self.data[['text','image_tensor']]
 
     def __len__(self):
